fix test2 bug
Fixed bug: unexpected error by checking if the pin count is == 0 before remove a frame from the hashmap. Otherwise there would be strange bug in the code.
*** Unexpected error(s) logged, test failed: First error occurred: --> DB Manager[from the <<Unknown>>]: /home/scratch/teaching/dsi-ht12/minibase/trunk/minibase-complete/spacemgr/db.cpp:609

I guess the reason is that if it removes from the database first, in test2 it called freePage() on this page as well, but it is considered  not in the buffers.